Move widget persistence and form tests into st_fragment_basics
These tests cover basic fragment behaviors (widget value persistence
across full reruns, form submission inside fragments) and belong in
the existing st_fragment_basics test suite rather than a separate file.

Also updates get_uuids and expect_only_fragment_uuid_changed to use
text-based markdown filtering instead of positional indexing, making
them robust to additional app scenarios.
