+++ e2e_playwright/st_fragment_basics_test.py
@@ -19,15 +19,21 @@ from e2e_playwright.conftest import wait_for_app_run
 from e2e_playwright.shared.app_utils import (
     click_button,
     click_checkbox,
+    click_form_button,
+    expect_markdown,
     select_selectbox_option,
 )
 
 
 def get_uuids(app: Page) -> tuple[str, str]:
-    expect(app.get_by_test_id("stMarkdown")).to_have_count(2)
+    markdowns = app.get_by_test_id("stMarkdown")
+    in_fragment = markdowns.filter(has_text="inside fragment:")
+    outside_fragment = markdowns.filter(has_text="outside: fragment")
+    expect(in_fragment).to_have_count(1)
+    expect(outside_fragment).to_have_count(1)
 
-    text_in_fragment = app.get_by_test_id("stMarkdown").first.text_content()
-    text_outside_fragment = app.get_by_test_id("stMarkdown").last.text_content()
+    text_in_fragment = in_fragment.text_content()
+    text_outside_fragment = outside_fragment.text_content()
 
     assert text_in_fragment is not None
     assert text_outside_fragment is not None
@@ -38,10 +44,11 @@ def get_uuids(app: Page) -> tuple[str, str]:
 def expect_only_fragment_uuid_changed(
     app: Page, old_text_in_fragment: str, old_text_outside_fragment: str
 ):
-    expect(app.get_by_test_id("stMarkdown").first).not_to_have_text(
+    markdowns = app.get_by_test_id("stMarkdown")
+    expect(markdowns.filter(has_text="inside fragment:")).not_to_have_text(
         old_text_in_fragment
     )
-    expect(app.get_by_test_id("stMarkdown").last).to_have_text(
+    expect(markdowns.filter(has_text="outside: fragment")).to_have_text(
         old_text_outside_fragment
     )
 
@@ -230,9 +237,54 @@ def test_full_app_rerun(app: Page):
     app.keyboard.press("r")
     wait_for_app_run(app)
 
-    expect(app.get_by_test_id("stMarkdown").first).not_to_have_text(
+    markdowns = app.get_by_test_id("stMarkdown")
+    expect(markdowns.filter(has_text="inside fragment:")).not_to_have_text(
         old_text_in_fragment
     )
-    expect(app.get_by_test_id("stMarkdown").last).not_to_have_text(
+    expect(markdowns.filter(has_text="outside: fragment")).not_to_have_text(
         old_text_outside_fragment
     )
+
+
+def test_fragment_widget_persists_across_full_app_rerun(app: Page):
+    """A widget inside a fragment retains its value after a full app rerun."""
+    slider = app.get_by_test_id("stSlider").get_by_role("slider")
+    slider.press("ArrowRight")
+    wait_for_app_run(app)
+
+    expect_markdown(app, "slider value: 51")
+
+    old_app_uuid = (
+        app.get_by_test_id("stMarkdown").filter(has_text="app uuid:").text_content()
+    )
+    assert old_app_uuid is not None
+
+    click_button(app, "Trigger full rerun")
+
+    expect(
+        app.get_by_test_id("stMarkdown").filter(has_text="app uuid:")
+    ).not_to_have_text(old_app_uuid)
+
+    expect_markdown(app, "slider value: 51")
+    expect(app.get_by_test_id("stSliderThumbValue").first).not_to_have_text("50")
+
+
+def test_form_inside_fragment_submits_correctly(app: Page):
+    """An st.form inside a fragment batches widget values and only applies
+    them on submit.
+    """
+    expect_markdown(app, "not submitted")
+
+    name_input = app.get_by_test_id("stTextInput").locator("input")
+    name_input.fill("Alice")
+    name_input.press("Enter")
+
+    expect_markdown(app, "not submitted")
+
+    click_form_button(app, "Submit form")
+
+    expect_markdown(app, "submitted: Alice")
+
+    expect(
+        app.get_by_test_id("stMarkdown").filter(has_text="submitted:")
+    ).to_have_count(1)
